Categorize enterprise frontend startup behaviors.
