+++ enterprise/cmd/frontend/internal/licensing/init/init.go
@@ -1,0 +1,69 @@
+package init
+
+import (
+	"context"
+
+	"github.com/sourcegraph/sourcegraph/cmd/frontend/enterprise"
+	"github.com/sourcegraph/sourcegraph/cmd/frontend/envvar"
+	"github.com/sourcegraph/sourcegraph/cmd/frontend/graphqlbackend"
+	"github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/dotcom/productsubscription"
+	"github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/licensing"
+	"github.com/sourcegraph/sourcegraph/internal/db"
+	"github.com/sourcegraph/sourcegraph/internal/goroutine"
+
+	_ "github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/auth"
+	_ "github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/graphqlbackend"
+	_ "github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/registry"
+)
+
+// TODO(efritz) - de-globalize assignments in this function
+// TODO(efritz) - refactor licensing packages - this is a huge mess!
+func Init(ctx context.Context, enterpriseServices *enterprise.Services) error {
+	// Enforce the license's max user count by preventing the creation of new users when the max is
+	// reached.
+	db.Users.PreCreateUser = licensing.NewPreCreateUserHook(&usersStore{})
+
+	// Make the Site.productSubscription.productNameWithBrand GraphQL field (and other places) use the
+	// proper product name.
+	graphqlbackend.GetProductNameWithBrand = licensing.ProductNameWithBrand
+
+	// Make the Site.productSubscription.actualUserCount and Site.productSubscription.actualUserCountDate
+	// GraphQL fields return the proper max user count and timestamp on the current license.
+	graphqlbackend.ActualUserCount = licensing.ActualUserCount
+	graphqlbackend.ActualUserCountDate = licensing.ActualUserCountDate
+
+	noLicenseMaximumAllowedUserCount := licensing.NoLicenseMaximumAllowedUserCount
+	graphqlbackend.NoLicenseMaximumAllowedUserCount = &noLicenseMaximumAllowedUserCount
+
+	noLicenseWarningUserCount := licensing.NoLicenseWarningUserCount
+	graphqlbackend.NoLicenseWarningUserCount = &noLicenseWarningUserCount
+
+	// Make the Site.productSubscription GraphQL field return the actual info about the product license,
+	// if any.
+	graphqlbackend.GetConfiguredProductLicenseInfo = func() (*graphqlbackend.ProductLicenseInfo, error) {
+		info, err := licensing.GetConfiguredProductLicenseInfo()
+		if info == nil || err != nil {
+			return nil, err
+		}
+		return &graphqlbackend.ProductLicenseInfo{
+			TagsValue:      info.Tags,
+			UserCountValue: info.UserCount,
+			ExpiresAtValue: info.ExpiresAt,
+		}, nil
+	}
+
+	goroutine.Go(func() {
+		licensing.StartMaxUserCount(&usersStore{})
+	})
+	if envvar.SourcegraphDotComMode() {
+		goroutine.Go(productsubscription.StartCheckForUpcomingLicenseExpirations)
+	}
+
+	return nil
+}
+
+type usersStore struct{}
+
+func (usersStore) Count(ctx context.Context) (int, error) {
+	return db.Users.Count(ctx, nil)
+}
